+++ userapps/hah/xap-xively/dscache.h
@@ -4,7 +4,7 @@
 #define __DSCACHE_H__
 
 struct datastream {
-  unsigned int feed;
+  unsigned long feed;
   unsigned int id;
   char *tag;
   float value;
@@ -13,7 +13,7 @@ struct datastream {
   char *unit;
 };
 
-void updateDatastream(unsigned int feed, unsigned int id, char *tag, float value, char *min, char *max, char *unit);
+void updateDatastream(unsigned long feed, unsigned int id, char *tag, float value, char *min, char *max, char *unit);
 void xivelyWebUpdate(int interval, void *userData);
 
 #endif
